requirements.txt
Django>=4.2,<5.0
google-generativeai
langchain-google-genai
langchain
langchain-community
chromadb
pypdf
python-dotenv
sentence-transformers
